remove zod from accessing env variables
